RFCT Only compute features that are later used
This avoids having to use gawk to manipulate TSV files
